Fixed issues with invalid sqrt values
There seems to be some error in fitting leading to the minor-axis exceeding the major-axis, which means the square-root will fail. For now, it seems to be functioning regardless. Also added some example images to include in the readme; next up is plotting cross-section and generating example plots for those.