Update Toroidal_Orbits.py
Wrote a function "fitTorusOrbit" which takes an outer radius, the number of divisions for the cross-section and our ellipse, and returns the orbit of best fit. It returns using the ellipse terminology, and the orbital parameter terminology.

Also added a testing function to calculate the area difference for a wide range of radii. Another function calculates the area difference on the interior vs the exterior, as two signed distances.