Changes to Circle Hull Difference + New Images
Added a "circle fitting" parameter to decide whether to use the circle distance, since it's results are still sub-optimal. Added a new function called "fit_desired_circle" which exclusively uses the circle area for optimization (no reliance on a desired cross-section).

Updated all the example plot images to show the x and y axes, as well as the center of mass for our cross-section.