Cleaned up Relative Motion Plot
Removed unnecessary pieces of "relative_motion_plot" and simplified some sections. Relative Motion now also prints out the 2d standard deviation for each circle-associated points (in cross-section X and Y) and plots the average for the cross-section of a single orbit. Also added grey lines for the x-axis (x = 0) and y = 1, to bring in line with multiorbit plotting format.