Update Toroidal_Orbits.py
Changed the functions to allow for outer_radii greater than 1, which produce self-intersecting toruses. This is done by modifying "generateTorusCrossSection" to take the absolute value of the x coordinate, since our cross-section always has a positive x value.
Also added a specific check for fitting, where it makes sure we do not take the arcsin of a value greater than 1 (which would return NaN)