Updated mult_orbit_density
Mult_Orbit_Density now displays not only angle => difference in angle, but also angle => arc area - desired area. This uses twin axes, for different data scales.

Also updated generatedEllipseEqArea to use the ParametricEllipse equation, to avoid having a separate copy of "get_pt" for it's particular usecase. This might cause issues in the future, because for a scalar value of theta it's return shape is [2,] instead of [2,1]